+++ BlueGardensEESystem/src/main/java/com/netbuilder/entitymanagers/ProductManager.java
@@ -19,7 +19,7 @@ public interface ProductManager {
 	public ArrayList<Product> findByName(String name);
 	public ArrayList<Product> findByPrice(double price);
 	public ArrayList<Product> findByOutStock();
-	public Product findById(long id);
+	public Product findById(int id);
 	public ArrayList<Product> findAll();
 	public ArrayList<Product> findByPriceLessThan(double price);
